refactor: 게시글 상세 경로 수정
